started adding themoviedatabase.org api. restructure of files
