+++ code/src/main/java/uk/co/conoregan/showrenamer/exception/ShowNotFoundException.java
@@ -1,0 +1,29 @@
+/*
+ * This file is part of ShowRenamer.
+ *
+ * ShowRenamer is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * ShowRenamer is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with ShowRenamer.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package uk.co.conoregan.showrenamer.exception;
+
+public class ShowNotFoundException extends Exception {
+
+    /**
+     * Constructor for Exception when show is not found in a search
+     * @param title of show searched
+     */
+    public ShowNotFoundException(String title) {
+        super("Could not find " + title + " in API...");
+    }
+}
